[FIX] settingview mvvm으로 변경(#77)
