+++ Genti_iOS/Genti_iOS/Presentation/SignIn/ViewModel/SignInViewModel.swift
@@ -47,17 +47,12 @@ final class SignInViewModel: ViewModel {
         case .genderSelect(let gender):
             self.state.gender = gender
         case .birthYearSelect:
-            if !self.state.firstTap {
-                self.state.firstTap = true
-            }
+            if !self.state.firstTap { self.state.firstTap = true }
             self.state.showPicker.toggle()
         case .completeButtonTap:
             Task { await postUserInformation() }
         case .backgroundTap:
-            if state.showPicker {
-                state.showPicker = false
-            }
-
+            if state.showPicker { state.showPicker = false }
         }
     }
     
@@ -66,20 +61,20 @@ final class SignInViewModel: ViewModel {
         do {
             state.isLoading = true
             try await signInUseCase.signIn(gender: state.gender, birthYear: state.birthYear)
-            EventLogManager.shared.logEvent(.completeInfoget)
-            // MARK: - 추후 수정
-            EventLogManager.shared.logEvent(.singIn(type: .apple))
-            EventLogManager.shared.addUserProperty(to: .userEmail(email: "테스트이메일입니다"))
             state.isLoading = false
             router.routeTo(.mainTab)
         } catch(let error) {
-            state.isLoading = false
-            guard let error = error as? GentiError else {
-                state.showAlert = .reportUnknownedError(error: error, action: { self.router.popToRoot() })
-                return
-            }
-            state.showAlert = .reportGentiError(error: error, action: nil)
+            self.handleError(error)
         }
+    }
+    
+    private func handleError(_ error: Error) {
+        state.isLoading = false
+        guard let error = error as? GentiError else {
+            state.showAlert = .reportUnknownedError(error: error, action: nil)
+            return
+        }
+        state.showAlert = .reportGentiError(error: error, action: nil)
     }
 }
 
@@ -88,42 +83,7 @@ extension SignInViewModel {
         return self.state.birthYear.formatterStyle(.none)
     }
     
-    func genderFont(_ gender: Gender) -> Font.PretendardType {
-        if gender == self.state.gender {
-            return .headline1
-        }
-        return .headline2
-    }
-    
-    func genderForegoundStyle(_ gender: Gender) -> Color {
-        if gender == self.state.gender {
-            return .green1
-        }
-        return .gray3
-    }
-    
     var isActive: Bool {
         return state.gender != nil && state.firstTap
     }
-    
-//    var completeButtonFont: Font.PretendardType {
-//        if isActive {
-//            return .headline1
-//        }
-//        return .headline2
-//    }
-//    
-//    var completeButtonBackground: Color {
-//        if isComplete {
-//            return .gentiGreen
-//        }
-//        return .gray6
-//    }
-//    
-//    var completeButtonForeground: Color {
-//        if isComplete {
-//            return .white
-//        }
-//        return .gray2
-//    }
 }
